feat: implement temporal file handling for rename operations and update related tests
